+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 This is the log of notable changes to the Design System Nimbus that are developer-facing.
 Package-specific changes not released in any package will be added here just before the release. Until then, you can find them in changelogs of the individual packages (see [packages](./packages) directory).
 
+## 2022-09-01
+
+### 📚 3rd party library updates
+
+- Added `@storybook/addon-controls@6.5.10`. ([#20](https://github.com/TiendaNube/nimbus-design-system/pull/20) by [@juniorconquista](https://github.com/juniorconquista))
+
 ## 2022-08-26
 
 ### 📚 3rd party library updates